requirements.txt
pyautogui
numpy
Pillow
